v2/requirements.txt
langchain
pinecone-client
langchain-pinecone
python-dotenv
pypdf 
sentence_transformers
google-generativeai
streamlit
